refactored
original script was making 2x requests to server due to `_loadScript` setting the `src` attribute of the `<script>` element ... as it turns out you can recursively call `requireScript` from `_cacheScript` and eat your own dog food.  This kills 2 birds with one stone, it doesnt double load content on the first pass, and it shrinks the total size of the script itself because `_loadScript` can be deleted

I also made the script AMD compatible

It should be noted that removing `_loadScript` breaks backwards compatibility for browsers that do not support local-storage though I plan on adding this functionality back into the `requireScript` function along with some code to specifically check if the browser supports local storage.